+++ changelog.md
@@ -4,7 +4,9 @@
 ### Version 4.6 - 2023.12.28
 
 * New: metabox styling rebuilt and improved with sidebar styling support
+* Edit: various improvements to the `Sanitize` class
 * Edit: expanded the `KSES` validation for the `Input` control
+* Fix: method for basic cleanup in `Sanitize` class fails in some cases
 * Fix: input rendering class stripping `min` and `max` attributes
 * Fix: checkbox and radio rendering not matching selected value